refactor: exported product data to a separate file
